requirements.txt
pynput
pystray
Pillow
pyinstaller
requests
python-dotenv
